create file for comment apis and refactor comment routes
