+++ src/screens/Project/TaskManagement/TicketDialog/Comments/CommentContent.tsx
@@ -1,5 +1,3 @@
-import { updateComment } from 'utils/api/tickets'
-
 export const CommentContent = ({ comment }) => {
   const { content } = comment
   return (
