+++ src/screens/Project/TaskManagement/TicketDialog/Comments/InputBanner.tsx
@@ -1,10 +1,10 @@
 import { InputAdornment, TextField } from '@mui/material'
 import { useState } from 'react'
-import { createComment } from 'utils/api/tickets'
 import SendOutlinedIcon from '@mui/icons-material/SendOutlined'
 import { CommentType } from 'interfaces/TaskBoardInterface'
 import { useAppSelector } from 'utils/redux/hooks'
 import { selectAuthUser } from 'utils/redux/slices/userSlice'
+import { createComment } from 'utils/api/comments'
 
 export const NewComment = ({
   commentType,
