+++ src/utils/api/tickets.ts
@@ -58,56 +58,3 @@ export const deleteTicketApi = async ticketData => {
     return { error: { status: 500, message: 'Ticket failed to delete' } }
   }
 }
-
-export const getTicketComments = async (
-  ticketId: number
-): Promise<Comment[]> => {
-  try {
-    const response = await api.get(`/ticket/${ticketId}/comments`)
-    return response.data.comments
-  } catch (err) {
-    console.error(err)
-    return []
-  }
-}
-
-// Comments
-export const createComment = async commentData => {
-  try {
-    const ticketData = await api.post('/createComment', commentData)
-    return ticketData.data
-  } catch (error) {
-    return { error: { status: 500, message: 'Failed to create comment' } }
-  }
-}
-
-export const deleteComment = async commentId => {
-  try {
-    const response = await api.delete(`/comments/${commentId}`)
-    return { status: response.status }
-  } catch (error) {
-    return { status: 500, message: 'Failed to delete comment' }
-  }
-}
-
-// BC: 761
-export const updateComment = async (commentId, commentUpdates) => {
-  try {
-    const response = await api.patch(
-      `/updateComment/${commentId}`,
-      commentUpdates
-    )
-    return { status: 200, ...response.data }
-  } catch (error) {
-    return { error: { status: 500, message: 'Failed to update comment' } }
-  }
-}
-
-export const getReplies = async (commentId: number): Promise<Comment[]> => {
-  try {
-    const response = await api.get(`/comment/${commentId}/replies`)
-    return response.data
-  } catch (error) {
-    return []
-  }
-}
